+++ core/src/plugins/filed/python/module/bareosfd.cc
@@ -759,6 +759,9 @@ static inline PyRestorePacket* NativeToPyRestorePacket(restore_pkt* rp)
 #else
     pRestorePacket->filedes = rp->filedes;
 #endif
+
+    pRestorePacket->original_file_name = rp->original_file_name;
+    pRestorePacket->original_link_name = rp->original_link_name;
   }
 
   return pRestorePacket;
@@ -2146,11 +2149,12 @@ static PyObject* PyRestorePacket_repr(PyRestorePacket* self)
        ", file_index=%" PRId32 ", linkFI=%" PRId32 ", uid=%" PRIu32
        ", statp=\"%s\", attrEx=\"%s\", ofname=\"%s\""
        ", olname=\"%s\", where=\"%s\", RegexWhere=\"%s\", replace=%d"
-       ", create_status=%d)",
+       ", create_status=%d, original_file_name=\"%s\""
+       ", original_link_name=\"%s\")",
        self->stream, self->data_stream, self->type, self->file_index,
        self->LinkFI, self->uid, PyGetStringValue(stat_repr), self->attrEx,
        self->ofname, self->olname, self->where, self->RegexWhere, self->replace,
-       self->create_status);
+       self->create_status, self->original_file_name, self->original_link_name);
 
   s = PyUnicode_FromString(buf.c_str());
   Py_DECREF(stat_repr);
@@ -2163,13 +2167,27 @@ static int PyRestorePacket_init(PyRestorePacket* self,
                                 PyObject* args,
                                 PyObject* kwds)
 {
-  static char* kwlist[]
-      = {(char*)"stream",     (char*)"data_stream",   (char*)"type",
-         (char*)"file_index", (char*)"linkFI",        (char*)"uid",
-         (char*)"statp",      (char*)"attrEX",        (char*)"ofname",
-         (char*)"olname",     (char*)"where",         (char*)"regexwhere",
-         (char*)"replace",    (char*)"create_status", NULL};
+  static char* kwlist[] = {(char*)"stream",
+                           (char*)"data_stream",
+                           (char*)"type",
+                           (char*)"file_index",
+                           (char*)"linkFI",
+                           (char*)"uid",
+                           (char*)"statp",
+                           (char*)"attrEX",
+                           (char*)"ofname",
+                           (char*)"olname",
+                           (char*)"where",
+                           (char*)"regexwhere",
+                           (char*)"replace",
+                           (char*)"create_status",
+                           (char*)"filedes",
+                           (char*)"original_file_name",
+                           (char*)"original_link_name",
+                           NULL};
 
+  // this has to be done like so, as we cannot overwrite the
+  // python object header
   self->stream = 0;
   self->data_stream = 0;
   self->type = 0;
@@ -2185,12 +2203,17 @@ static int PyRestorePacket_init(PyRestorePacket* self,
   self->replace = 0;
   self->create_status = 0;
 
+  self->filedes = kInvalidFiledescriptor;
+  self->original_file_name = nullptr;
+  self->original_link_name = nullptr;
+
   if (!PyArg_ParseTupleAndKeywords(
           args, kwds, "|iiiiiIosssssii", kwlist, &self->stream,
           &self->data_stream, &self->type, &self->file_index, &self->LinkFI,
           &self->uid, &self->statp, &self->attrEx, &self->ofname, &self->olname,
-          &self->where, &self->RegexWhere, &self->replace,
-          &self->create_status)) {
+          &self->where, &self->RegexWhere, &self->replace, &self->create_status,
+          &self->filedes, &self->original_file_name,
+          &self->original_link_name)) {
     return -1;
   }